+++ changelog.md
@@ -2,8 +2,8 @@
 
 ## v0.4 (forthcoming)
 
+  * Added options to download and import the v2.2016 (original v2) and v3.2026 (original v3) datasets.
   * Switched to [Pooch](https://github.com/fatiando/pooch) for download handling.
-  * Added option to download and import the original v2.2016 dataset.
 
 ## v0.3 (26 May 2025)